feat(launch file refactoring): moved common functions to common_functions_launch file inside reseq_ros2
