+++ reseq_ros2/launch/digital_twin_launch.py
@@ -7,33 +7,7 @@ from yaml import SafeLoader
 from launch_ros.actions import Node
 from ament_index_python.packages import get_package_share_directory
 import xacro
-
-#Default config file path
-share_folder = get_package_share_directory("reseq_ros2")
-config_path = f'{share_folder}/config'
-default_filename = "reseq_mk1_can.yaml"
-
-def parse_config(filename):
-    with open(filename) as f:
-        return yaml.load(f, Loader=SafeLoader)
-    
-def get_addresses(config):
-    l = []
-    for mod in config['modules']:
-        l.append(mod['address'])
-    return l
-
-def get_joints(config):
-    l = []
-    for mod in config['modules']:
-        if mod['hasJoint']:
-            l.append(mod['address'])
-    return l
-
-def get_end_effector(config):
-    for mod in config['modules']:
-        if mod['hasEndEffector']:
-            return mod['address']
+from reseq_ros2.common_functions_launch import *
 
 #launch_setup is used through an OpaqueFunction because it is the only way to manipulate a command line argument directly in the launch file
 def launch_setup(context, *args, **kwargs):
